refactor: move createAnswer to _lib for better structure
